Attempting to make compatible with Python 2 again.
